Move player_interaction_system into 'player' directory
